+++ packages/shared/src/hooks/collectionDataContext.d.ts
@@ -1,7 +1,0 @@
-import type { CollectionDataServices } from "../types";
-export declare const CollectionDataContext: import("react").Context<CollectionDataServices>;
-/**
- * Collection Data 서비스 훅
- */
-export declare function useCollectionDataServices(): CollectionDataServices;
-//# sourceMappingURL=collectionDataContext.d.ts.map