+++ packages/shared/src/runtime/index.d.ts
@@ -1,8 +1,0 @@
-/**
- * Runtime Index
- *
- * @since 2026-01-02 Phase 3
- */
-export { ActionExecutor } from './ActionExecutor';
-export { PublishEventRuntime } from './PublishEventRuntime';
-//# sourceMappingURL=index.d.ts.map